# Conflicts:
#	crates/backend-uzu/src/backends/common/kernel/quant_matmul.rs
#	crates/backend-uzu/src/backends/cpu/kernel/mod.rs
#	crates/backend-uzu/src/backends/cpu/kernel/quant_matmul/mod.rs
#	crates/backend-uzu/src/backends/cpu/kernel/quant_matmul/qmm_transposed.rs
#	crates/backend-uzu/src/backends/cpu/kernel/quant_matmul/qmv_fast.rs
#	crates/backend-uzu/src/backends/cpu/mod.rs
#	crates/backend-uzu/src/backends/metal/kernel/mod.rs
#	crates/backend-uzu/src/backends/metal/kernel/quant_matmul/qmm_transposed.metal
#	crates/backend-uzu/src/backends/metal/kernel/quant_matmul/qmv_fast.metal
#	crates/backend-uzu/src/backends/metal/kernel/quant_matmul/quant_matmul.h
#	crates/backend-uzu/src/backends/metal/mod.rs
#	crates/backend-uzu/tests/performance/matmul/mod.rs
#	crates/backend-uzu/tests/unit/kernel/quant_matmul/mod.rs
#	crates/backend-uzu/tests/unit/kernel/quant_matmul/qmm_transposed_test.rs
#	crates/backend-uzu/tests/unit/kernel/quant_matmul/qmv_fast_test.rs

+++ crates/backend-remote/src/openai/bridging/chat/request.rs
@@ -45,7 +45,7 @@ pub fn build(
         request.reasoning_effort = Some(effort);
     }
     if let Some(token_limit) = config.token_limit {
-        request.max_completion_tokens = Some(token_limit as u32);
+        request.max_completion_tokens = Some(token_limit);
     }
     match &config.sampling_policy {
         SamplingPolicy::Default {} => {},
